requirements.txt
pyyaml==6.0.1
python-telegram-bot==20.7
python-telegram-bot[job-queue]
psutil==5.9.8